requirements.txt
asgiref==3.8.1
black==24.4.2
click==8.1.7
Django==4.2.4
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
gunicorn
zappa
djangorestframework==3.15.2
mangum
aws-sam-cli